agentic-platform/requirements.txt
# Gateway
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pyjwt[crypto]>=2.8.0
python-multipart>=0.0.6
httpx>=0.27.0

# Slack bot
slack-bolt>=1.18.0
slack-sdk>=3.27.0

# Admin UI
streamlit>=1.31.0

# Database seeding
mariadb>=1.1.10

# Embeddings & memory
# (OpenAI API called via httpx — no SDK dependency needed)

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
